SER-495 changed routing in image-review
