+++ front/main/app/models/image-review/summary.js
@@ -1,7 +1,7 @@
 import Ember from 'ember';
 import moment from 'moment';
 import request from 'ember-ajax/request';
-import ImageReviewItemModel from './image-review-item';
+import ImageReviewItemModel from './item';
 
 const ImageReviewSummaryModel = Ember.Object.extend({
 	summary: null,
