+++ front/main/app/routes/image-review.js
@@ -1,2 +1,10 @@
 import Ember from 'ember';
-export default Ember.Route.extend({});
+
+const {Route, Logger} = Ember;
+
+export default Route.extend({
+	beforeModel() {
+		Logger.error('I\'m in main route!');
+		this.transitionTo('image-review.index');
+	}
+})
