+++ front/main/app/templates/image-review/batch-id.hbs
@@ -2,7 +2,7 @@
 	model=model
 	reviewAndGetMoreImages=(action 'reviewAndGetMoreImages')
 	changeItemModel=(action 'changeItemModel')
-	openMainPage=(action 'openMainPage')
+	goBackToLastBatch=(action 'goBackToLastBatch')
 	openSummary=(action 'openSummary')
 	getAllWithStatus=(action 'getAllWithStatus')
 }}
